clean up namings and add comments to some tests
