+++ src/include/optimizer/logical_operator_visitor.h
@@ -33,7 +33,7 @@ protected:
         return op;
     }
 
-    virtual void visitScanNodeProperty(planner::LogicalOperator* /*op*/) {}
+    virtual void visitScanNodeTable(planner::LogicalOperator* /*op*/) {}
     virtual std::shared_ptr<planner::LogicalOperator> visitScanNodePropertyReplace(
         std::shared_ptr<planner::LogicalOperator> op) {
         return op;
