+++ src/include/planner/operator/logical_operator.h
@@ -51,7 +51,7 @@ enum class LogicalOperatorType : uint8_t {
     RECURSIVE_EXTEND,
     SCAN_FILE,
     SCAN_FRONTIER,
-    SCAN_NODE_PROPERTY,
+    SCAN_NODE_TABLE,
     SEMI_MASKER,
     SET_NODE_PROPERTY,
     SET_REL_PROPERTY,
