+++ src/include/planner/operator/logical_plan_util.h
@@ -25,7 +25,7 @@ private:
     static void encodeHashJoin(LogicalOperator* logicalOperator, std::string& encodeString);
     static void encodeExtend(LogicalOperator* logicalOperator, std::string& encodeString);
     static void encodeRecursiveExtend(LogicalOperator* logicalOperator, std::string& encodeString);
-    static void encodeScanNodeProperty(LogicalOperator* logicalOperator, std::string& encodeString);
+    static void encodeScanNodeTable(LogicalOperator* logicalOperator, std::string& encodeString);
 };
 
 } // namespace planner
