+++ src/include/processor/operator/scan/scan_table.h
@@ -14,7 +14,7 @@ public:
         : PhysicalOperator{operatorType, std::move(child), id, paramString}, nodeIDPos{nodeIDPos},
           outVectorsPos{std::move(outVectorsPos)}, nodeIDVector{nullptr} {}
 
-    // Scan Node Table constructor.
+    // ScanNodeTable is a source operator without a child op.
     ScanTable(PhysicalOperatorType operatorType, const DataPos& nodeIDPos,
         std::vector<DataPos> outVectorsPos, uint32_t id, const std::string& paramString)
         : PhysicalOperator{operatorType, id, paramString}, nodeIDPos{nodeIDPos},
