+++ src/planner/join_order/cardinality_estimator.cpp
@@ -2,7 +2,7 @@
 
 #include "binder/expression/property_expression.h"
 #include "planner/join_order/join_order_util.h"
-#include "planner/operator/scan/logical_scan_node_property.h"
+#include "planner/operator/scan/logical_scan_node_table.h"
 
 using namespace kuzu::binder;
 using namespace kuzu::common;
@@ -34,7 +34,7 @@ void CardinalityEstimator::addNodeIDDom(const binder::Expression& nodeID,
 }
 
 uint64_t CardinalityEstimator::estimateScanNode(LogicalOperator* op) {
-    auto scan = ku_dynamic_cast<LogicalOperator*, LogicalScanNodeProperty*>(op);
+    auto scan = ku_dynamic_cast<LogicalOperator*, LogicalScanNodeTable*>(op);
     return atLeastOne(getNodeIDDom(scan->getNodeID()->getUniqueName()));
 }
