+++ src/planner/operator/logical_operator.cpp
@@ -92,8 +92,8 @@ std::string LogicalOperatorUtils::logicalOperatorTypeToString(LogicalOperatorTyp
         return "SCAN_FILE";
     case LogicalOperatorType::SCAN_FRONTIER:
         return "SCAN_FRONTIER";
-    case LogicalOperatorType::SCAN_NODE_PROPERTY:
-        return "SCAN_NODE_PROPERTY";
+    case LogicalOperatorType::SCAN_NODE_TABLE:
+        return "SCAN_NODE_TABLE";
     case LogicalOperatorType::SEMI_MASKER:
         return "SEMI_MASKER";
     case LogicalOperatorType::SET_NODE_PROPERTY:
