+++ src/planner/operator/logical_plan_util.cpp
@@ -5,7 +5,7 @@
 #include "planner/operator/extend/logical_recursive_extend.h"
 #include "planner/operator/logical_hash_join.h"
 #include "planner/operator/logical_intersect.h"
-#include <planner/operator/scan/logical_scan_node_property.h>
+#include "planner/operator/scan/logical_scan_node_table.h"
 
 using namespace kuzu::binder;
 
@@ -47,8 +47,8 @@ void LogicalPlanUtil::encodeJoinRecursive(LogicalOperator* logicalOperator,
         encodeRecursiveExtend(logicalOperator, encodeString);
         encodeJoinRecursive(logicalOperator->getChild(0).get(), encodeString);
     } break;
-    case LogicalOperatorType::SCAN_NODE_PROPERTY: {
-        encodeScanNodeProperty(logicalOperator, encodeString);
+    case LogicalOperatorType::SCAN_NODE_TABLE: {
+        encodeScanNodeTable(logicalOperator, encodeString);
     } break;
     default:
         for (auto i = 0u; i < logicalOperator->getNumChildren(); ++i) {
@@ -86,10 +86,9 @@ void LogicalPlanUtil::encodeRecursiveExtend(LogicalOperator* logicalOperator,
     encodeString += "RE(" + logicalExtend->getNbrNode()->toString() + ")";
 }
 
-void LogicalPlanUtil::encodeScanNodeProperty(LogicalOperator* logicalOperator,
+void LogicalPlanUtil::encodeScanNodeTable(LogicalOperator* logicalOperator,
     std::string& encodeString) {
-    auto scan =
-        common::ku_dynamic_cast<LogicalOperator*, LogicalScanNodeProperty*>(logicalOperator);
+    auto scan = common::ku_dynamic_cast<LogicalOperator*, LogicalScanNodeTable*>(logicalOperator);
     encodeString += "S(" + scan->getNodeID()->toString() + ")";
 }
