+++ src/planner/plan/CMakeLists.txt
@@ -21,7 +21,7 @@ add_library(kuzu_planner_plan_operator
         append_order_by.cpp
         append_projection.cpp
         append_scan_file.cpp
-        append_scan_node.cpp
+        append_scan_node_table.cpp
         append_set.cpp
         append_simple.cpp
         append_unwind.cpp
