+++ src/planner/plan/append_scan_node_table.cpp
@@ -1,5 +1,5 @@
 #include "binder/expression/property_expression.h"
-#include "planner/operator/scan/logical_scan_node_property.h"
+#include "planner/operator/scan/logical_scan_node_table.h"
 #include "planner/planner.h"
 
 using namespace kuzu::common;
@@ -8,17 +8,16 @@ using namespace kuzu::binder;
 namespace kuzu {
 namespace planner {
 
-void Planner::appendScanNodeProperties(std::shared_ptr<Expression> nodeID,
+void Planner::appendScanNodeTable(std::shared_ptr<Expression> nodeID,
     std::vector<table_id_t> tableIDs, const expression_vector& properties, LogicalPlan& plan) {
     expression_vector propertiesToScan_;
     for (auto& property : properties) {
-        // TODO: Double check if this is necessary. If so, better move out of here.
         if (ku_dynamic_cast<Expression&, PropertyExpression&>(*property).isInternalID()) {
             continue;
         }
         propertiesToScan_.push_back(property);
     }
-    auto scanNodeProperty = make_shared<LogicalScanNodeProperty>(std::move(nodeID),
+    auto scanNodeProperty = make_shared<LogicalScanNodeTable>(std::move(nodeID),
         std::move(tableIDs), propertiesToScan_);
     scanNodeProperty->computeFactorizedSchema();
     plan.setLastOperator(std::move(scanNodeProperty));
