+++ src/processor/operator/scan/CMakeLists.txt
@@ -1,7 +1,7 @@
 add_library(kuzu_processor_operator_scan
         OBJECT
-        scan_multi_node_tables.cpp
         scan_multi_rel_tables.cpp
+        scan_node_table.cpp
         scan_rel_table.cpp
         scan_table.cpp)
