+++ src/processor/operator/scan/scan_node_table.cpp
@@ -1,48 +1,75 @@
 #include "processor/operator/scan/scan_node_table.h"
 
 using namespace kuzu::common;
-using namespace kuzu::transaction;
 
 namespace kuzu {
 namespace processor {
 
-// void ScanNodeTableSharedState::initialize(const Transaction* transaction,
-//     storage::NodeTable* table) {
-//     this->table = table;
-//     this->currentGroupIdx = 0;
-//     this->numNodeGroups = table->getNumNodeGroups(transaction);
-// }
-//
-// node_group_idx_t ScanNodeTableSharedState::getNextMorsel() {
-//     std::unique_lock lck{mtx};
-//     if (currentGroupIdx == numNodeGroups) {
-//         return INVALID_NODE_GROUP_IDX;
-//     }
-//     return currentGroupIdx++;
-// }
-//
-// void ScanNodeTable::initLocalStateInternal(ResultSet* resultSet, ExecutionContext* context) {
-//     ScanTable::initLocalStateInternal(resultSet, context);
-//     readState = std::make_unique<storage::NodeTableReadState>(info->columnIDs);
-//     initVectors(*readState, *resultSet);
-// }
-//
-// bool ScanNodeTable::getNextTuplesInternal(ExecutionContext* context) {
-//     do {
-//         if (readState->hasMoreToRead(context->clientContext->getTx())) {
-//             info->table->read(context->clientContext->getTx(), *readState);
-//             return true;
-//         }
-//         const auto nodeGroupIdx = sharedState->getNextMorsel();
-//         if (nodeGroupIdx == INVALID_NODE_GROUP_IDX) {
-//             return false;
-//         }
-//         info->table->initializeReadState(context->clientContext->getTx(), nodeGroupIdx,
-//             info->columnIDs, *readState);
-//         info->table->read(context->clientContext->getTx(), *readState);
-//     } while (readState->nodeIDVector->state->getSelVector().getSelSize() == 0);
-//     return true;
-// }
+void ScanNodeTableSharedState::initialize(const transaction::Transaction* transaction,
+    storage::NodeTable* table_) {
+    this->table = table_;
+    this->currentGroupIdx = 0;
+    this->numNodeGroups = table->getNumNodeGroups(transaction);
+}
+
+node_group_idx_t ScanNodeTableSharedState::getNextMorsel() {
+    std::unique_lock lck{mtx};
+    if (currentGroupIdx == numNodeGroups) {
+        return INVALID_NODE_GROUP_IDX;
+    }
+    return currentGroupIdx++;
+}
+
+void ScanNodeTables::initLocalStateInternal(ResultSet* resultSet, ExecutionContext* context) {
+    ScanTable::initLocalStateInternal(resultSet, context);
+    for (const auto& info : infos) {
+        auto readState = std::make_unique<storage::NodeTableReadState>(info->columnIDs);
+        initVectors(*readState, *resultSet);
+        readStates.push_back(std::move(readState));
+    }
+}
+
+void ScanNodeTables::initGlobalStateInternal(ExecutionContext* context) {
+    KU_ASSERT(sharedStates.size() == infos.size());
+    for (auto i = 0u; i < infos.size(); i++) {
+        sharedStates[i]->initialize(context->clientContext->getTx(), infos[i]->table);
+    }
+}
+
+bool ScanNodeTables::getNextTuplesInternal(ExecutionContext* context) {
+    while (currentTableIdx < infos.size()) {
+        const auto info = infos[currentTableIdx].get();
+        const auto readState = readStates[currentTableIdx].get();
+        while (true) {
+            if (readState->hasMoreToRead(context->clientContext->getTx())) {
+                info->table->read(context->clientContext->getTx(), *readState);
+                return true;
+            }
+            const auto nodeGroupIdx = sharedStates[currentTableIdx]->getNextMorsel();
+            if (nodeGroupIdx == INVALID_NODE_GROUP_IDX) {
+                // Current table out of data, move to the next one.
+                currentTableIdx++;
+                break;
+            }
+            info->table->initializeReadState(context->clientContext->getTx(), nodeGroupIdx,
+                info->columnIDs, *readState);
+            info->table->read(context->clientContext->getTx(), *readState);
+            if (readState->nodeIDVector->state->getSelVector().getSelSize() > 0) {
+                return true;
+            }
+        }
+    }
+    return false;
+}
+
+std::unique_ptr<PhysicalOperator> ScanNodeTables::clone() {
+    std::vector<std::unique_ptr<ScanNodeTableInfo>> clonedInfos;
+    for (const auto& info : infos) {
+        clonedInfos.push_back(info->copy());
+    }
+    return make_unique<ScanNodeTables>(nodeIDPos, outVectorsPos, std::move(clonedInfos),
+        sharedStates, id, paramsString);
+}
 
 } // namespace processor
 } // namespace kuzu
